Move result types to @proti-iac/core
