+++ proti-core/src/index.ts
@@ -6,6 +6,7 @@ export * from './generator';
 export * from './pulumi-project';
 export * from './oracle';
 export * from './plugin';
+export * from './result';
 export * from './spec-impl';
 export * from './test-coordinator';
 export * from './utils';
