+++ proti-reporter/package.json
@@ -27,7 +27,6 @@
   },
   "devDependencies": {
     "@proti-iac/core": "workspace:^",
-    "@proti-iac/test-runner": "workspace:^",
     "@types/jest": "^29.5.7",
     "@types/node": "^18.18.8",
     "jest": "^29.7.0",
